wip: syntax for the 'lang' keyword and new AST objects
This is a little broken but it is the beginning of what the
new syntax will look like with a keyword that allows defining
different languages in the same Module.  Speaking of which,
this change also includes Module as a new type of AST object
to be used as a container for lang entries and to be mapped
to files in the file system.